move context-enabled helpers to separate folders
