+++ samples/test/steps/when.ts
@@ -31,12 +31,12 @@ import confirmValidPassword from '../support/action/confirmValidPassword';
 // import setInputField from '../support/action/setInputField';
 // import setPromptText from '../support/action/setPromptText';
 
-import enterCredential from '../support/action/enterCredential';
+import enterCredential from '../support/action/context-enabled/enterCredential';
 import enterValidPassword from '../support/action/enterValidPassword';
-import enterCorrectCode from '../support/action/live-user/enterCorrectCode';
-import enterLiveUserEmail from '../support/action/live-user/enterEmail';
+import enterCorrectCode from '../support/action/context-enabled/live-user/enterCorrectCode';
+import enterLiveUserEmail from '../support/action/context-enabled/live-user/enterEmail';
 import submitAnyForm from '../support/action/submitAnyForm';
-import submitForm from '../support/action/submitForm';
+import submitForm from '../support/action/context-enabled/submitForm';
 import clickLogout from '../support/action/clickLogout';
 import submitPasswordRecoverForm from '../support/action/submitPasswordRecoverForm';
 import selectAuthenticator from '../support/action/selectAuthenticator';
@@ -45,26 +45,26 @@ import selectGoogleAuthenticator from '../support/action/selectGoogleAuthenticat
 import selectSecurityQuestionAuthenticator from '../support/action/selectSecurityQuestionAuthenticator';
 import enterIncorrectCode from '../support/action/enterIncorrectCode';
 import inputInvalidEmail from '../support/action/inputInvalidEmail';
-import enterRegistrationField from '../support/action/live-user/enterRegistrationField';
+import enterRegistrationField from '../support/action/context-enabled/live-user/enterRegistrationField';
 import selectPasswordAuthenticator from '../support/action/selectPasswordAuthenticator';
 import selectPhoneAuthenticator from '../support/action/selectPhoneAuthenticator';
-import enterCorrectSMSCode from '../support/action/live-user/enterCorrectSMSCode';
+import enterCorrectSMSCode from '../support/action/context-enabled/live-user/enterCorrectSMSCode';
 import selectSmsAuthenticator from '../support/action/selectSmsAuthenticator';
-import enterCorrectPhoneNumber from '../support/action/live-user/enterCorrectPhoneNumber';
+import enterCorrectPhoneNumber from '../support/action/context-enabled/live-user/enterCorrectPhoneNumber';
 import selectVerifyBySms from '../support/action/selectVerifyBySms';
 import skipForm from '../support/action/skipForm';
 import inputInvalidEmailFormat from '../support/action/inputInvalidEmailFormat';
 import enterIncorrectPhoneNumberFormat from '../support/action/enterIncorrectPhoneNumberFormat';
 import clickFacebookButton from '../support/action/clickFacebookButton';
 import clickLoginWithFacebookInWidget from '../support/action/clickLoginWithFacebookInWidget';
-import signInIntoFacebook from '../support/action/signInIntoFacebook';
+import signInIntoFacebook from '../support/action/context-enabled/signInIntoFacebook';
 import clickLoginWithOktaOIDCIdPInWidget from '../support/action/clickLoginWithOktaOIDCIdPInWidget';
 import clickOIDCIdPButton from '../support/action/clickOIDCIdPButton';
 import loginWidget from '../support/action/loginWidget';
 import enterCorrectQuestionAnswer from '../support/action/enterCorrectQuestionAnswer';
-import scanQrCode from '../support/action/scanQrCode';
-import getSharedSecret from '../support/action/getSharedSecret';
-import enterCorrectGoogleAuthenticatorCode from '../support/action/enterCorrectGoogleAuthenticatorCode';
+import scanQrCode from '../support/action/context-enabled/scanQrCode';
+import getSharedSecret from '../support/action/context-enabled/getSharedSecret';
+import enterCorrectGoogleAuthenticatorCode from '../support/action/context-enabled/enterCorrectGoogleAuthenticatorCode';
 import ActionContext from '../support/context';
 
 When(
