+++ samples/test/support/action/context-enabled/live-user/enterCorrectCode.ts
@@ -11,9 +11,12 @@
  */
 
 
-import setInputField from '../setInputField';
-import ActionContext from '../../context';
+import ChallengeAuthenticator from '../../../selectors/ChallengeAuthenticator';
+import a18nClient from '../../../management-api/a18nClient';
+import setInputField from '../../setInputField';
+import ActionContext from '../../../context';
 
 export default async function (this: ActionContext) {
-  await setInputField('set', this.credentials.password, '#password');
+  const code = await a18nClient.getEmailCode(this.credentials.profileId);
+  await setInputField('set', code, ChallengeAuthenticator.code);
 }
