+++ samples/test/support/action/context-enabled/live-user/enterEmail.ts
@@ -11,10 +11,10 @@
  */
 
 
-import EnrollPhoneAuthenticator from '../../selectors/EnrollPhoneAuthenticator';
-import setInputField from '../setInputField';
-import ActionContext from '../../context';
+import setInputField from '../../setInputField';
+import PasswordRecover from '../../../selectors/PasswordRecover';
+import ActionContext from '../../../context';
 
-export default async function (this: ActionContext) {
-  await setInputField('set', this.credentials.phoneNumber, EnrollPhoneAuthenticator.phoneNumber);
+export default async function(this: ActionContext) {
+  await setInputField('set', this.user.profile.email, PasswordRecover.username);
 }
