+++ samples/test/support/action/context-enabled/live-user/navigateToLoginAndAuthenticate.ts
@@ -11,11 +11,17 @@
  */
 
 
-import EnrollGoogleAuthenticator from '../selectors/EnrollGoogleAuthenticator';
-import getText from './getText';
-import ActionContext from '../context';
+import navigateTo from '../../navigateTo';
+import loginDirect from '../../loginDirect';
+import ActionContext from '../../../context';
 
-export default async function (this: ActionContext) {
-  const sharedSecret = await getText( EnrollGoogleAuthenticator.sharedSecret );
-  this.sharedSecret = sharedSecret;
+export default async function (
+    this: ActionContext,
+    userName: string
+) {
+    await navigateTo(userName, 'Login with Username and Password');
+    await loginDirect({
+        username: this.credentials?.emailAddress,
+        password: this.credentials?.password
+    });
 }
