+++ samples/test/support/action/context-enabled/org-config/deleteTestPolicies.ts
@@ -1,5 +1,5 @@
-import ActionContext from '../../context';
-import deletePolicy from '../../management-api/deletePolicy';
+import ActionContext from '../../../context';
+import deletePolicy from '../../../management-api/deletePolicy';
 
 export default async function (this: ActionContext) {
   if (this.useProfileEnrollPolicy) {
