+++ samples/test/support/check/context-enabled/checkProfileName.ts
@@ -11,9 +11,9 @@
  */
 
 
-import checkEqualsText from './checkEqualsText';
-import UserHome from '../selectors/UserHome';
-import ActionContext from '../context';
+import checkEqualsText from '../checkEqualsText';
+import UserHome from '../../selectors/UserHome';
+import ActionContext from '../../context';
 
 export default async function(this: ActionContext) {
   const firstName = this.credentials?.firstName;
